Added a timer floating around the player. Timer stops after hitting red, ending tile.

+++ consts.hpp
@@ -1,10 +1,10 @@
 #ifndef PROJEKT_CONSTS_HPP
 #define PROJEKT_CONSTS_HPP
 
-constexpr unsigned short WINDOW_HEIGHT = 600;
-constexpr unsigned short WINDOW_WIDTH = 800;
+constexpr unsigned int WINDOW_HEIGHT = 600;
+constexpr unsigned int WINDOW_WIDTH = 800;
 constexpr float TILE = 80.0f;
-constexpr unsigned short MAP_WIDTH = 10;
-constexpr unsigned short MAP_HEIGHT = MAP_WIDTH;
+constexpr unsigned int MAP_WIDTH = 10;
+constexpr unsigned int MAP_HEIGHT = MAP_WIDTH;
 
 #endif //PROJEKT_CONSTS_HPP
